requirements.txt
aiogram==3.20.0
openai>=1.84.0
python-dotenv==1.1.0
pydantic-settings==2.9.1
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.3
amplitude-analytics==1.1.5
requests==2.32.3
redis==5.1.1